+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+	- README enhanced for non-HTML tags like <bid> <description>.
+	- Variable suits also correctly translated for their bids.
+	- bss2bml utility added.
+	
 ## [2.2.1] - 2021-04-15
 
 ### Changed